REFACTORING - int, long, ... durch uint8_t, uint16_t, uint32_t ersetzt (und ohne unsigned) - Ventilplatine nun auch in Steuerung komplett eingebunden - Bugs gefixt (und hoffentlich keine neuen eingebaut ........ ) - README.md aktualisiert

+++ controller/src/eventElement.h
@@ -3,8 +3,8 @@
 
 namespace control {
     typedef struct eventElementStruct {
-        int value;
-        unsigned long time;
+        int16_t value;
+        uint32_t time;
     } eventElement;
 
 }
